requirements.txt
Django==3.0.6
djangorestframework==3.11.0
#psycopg2==2.8.5
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
